requirements.txt
absl-py==0.12.0
aiohttp==3.7.4
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
gql==3.0.0a5
graphql-core==3.1.3
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
pathspec==0.8.1
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
